fix(httpz): update api
see https://github.com/oscartbeaumont/rspc/commit/e11e19d83077b8d139073bd9541f2f7c01334a10
