+++ Cargo.toml
@@ -15,7 +15,7 @@ categories = ["web-programming", "asynchronous"]
 default = []
 tauri = ["dep:tauri"]
 tracing = ["dep:tracing"]
-httpz = ["dep:httpz"]
+httpz = ["dep:httpz", "httpz/cookies"]
 # openapi = ["dep:openapiv3"]
 # playground = []
 
@@ -25,8 +25,8 @@ actix-web = ["httpz/actix-web", "httpz/tokio-ws", "httpz/axum"]
 # poem = ["httpz/poem"]
 rocket = ["httpz/rocket", "httpz/tokio-ws", "httpz/axum"]
 # warp = ["httpz/warp"]
-lambda = ["httpz/lambda", "httpz/tokio-ws", "httpz/axum"]
-workers = ["httpz/workers"]
+lambda = ["httpz/lambda", "httpz/ws", "httpz/axum"]
+workers = ["httpz/workers", "httpz/ws"]
 
 # Specta
 uuid = ["specta/uuid"]
